Zrní - Hákal
This is the latest version of my program. It displays syllables of the text while beeping the molody; however, when I uncomment the guitar solo in both the Melody and the Text class, the text and melody isn't synchronized. I would like to kindly ask you to help me. Thank you very much in advance.